Some final touches

+++ db/migrate/20230904213938_create_recipes.rb
@@ -2,8 +2,8 @@ class CreateRecipes < ActiveRecord::Migration[7.0]
   def change
     create_table :recipes do |t|
       t.string :name
-      t.integer :preparation_time
-      t.integer :cooking_time
+      t.decimal :preparation_time
+      t.decimal :cooking_time
       t.text :description
       t.boolean :is_public
       t.references :user, null: false, foreign_key: true
